Replace Result returns in API with callback based error handling
With this change, all functions and operations return appropriate
values/objects instead of wrapping those values inside a Result<>
object. All errors reported by FFI calls are handled by a global
error handling function that is registered by a call to the function
`register_error_handler` that accepts either a closure or a function.

By default, the error handler is set to a internal function that
prints the error message to standard output and then panics.