Fix for OSX platform w.r.t static arrays
For some reason, directly passing a `&[]` to function that is using
OpenCL backend is causing a SEGFAULT. This problem is however is
not happening with CPU or CUDA backend. Moving the slice to a independent
variable and then borrowing a reference is executing as expected on OSX.
This may have to do with slices and the memory they are pointing to, but it
is quite puzzling why it only happens on one of three backends.