Fix output array for logical ops AND, OR and NEQ
Earlier, these functions were using binary_func macro which picks
output based on the input arrays types. However, as these are logical
operations which always result in boolean output, the correct macro to
use for this is overloaded_logic_func.

With this fix, ::and, ::or and ::neq functions have acquired an
additional feature i.e. the user can also use scalars as one of the
inputs of these functions.