Fix logic to sequence based indexing in row/col/slice functions
Prior to this change, the following functions were not checking for additional
dimensional data beyond the dimension concerned with the particular
function.

- row
- col
- slice
- rows
- cols
- slices
- set_row
- set_col
- slice
- set_rows
- set_cols
- set_slices

Similar logic was missing in one particular matching pattern of view
macro which is also fixed in this change.

Few additional unit tests are added in macro and index module checking
for the pitfalls this change addresses